+++ src/shapes/ActiveSelection.ts
@@ -1,4 +1,4 @@
-import { ControlRenderingStyleOverride } from '../controls/controls.render';
+import type { ControlRenderingStyleOverride } from '../controls/controlRendering';
 import { TClassProperties } from '../typedefs';
 import { classRegistry } from '../util/class_registry';
 import { Group } from './Group';
